Modify Point.java to have x and y as doubles instead of ints
Better suits the given specification.

+++ app/src/main/java/decide/utils/Point.java
@@ -1,11 +1,13 @@
 package decide.utils;
 
+import java.lang.Math;
+
 // immutable standard Point class
 public final class Point {
-    final public int x;
-    final public int y;
+    final public double x;
+    final public double y;
 
-    public Point(int x, int y) {
+    public Point(double x, double y) {
         this.x = x;
         this.y = y;
     }
